+++ overrides/kubejs/client_scripts/99-tooltips.js
@@ -1,0 +1,5 @@
+function itemTooltips(ev) {
+    ev.add("ae2:spatial_anchor", Text.of("Broken on Fabric!").red());
+}
+
+onEvent("item.tooltip", itemTooltips);