+++ overrides/kubejs/server_scripts/30-tags.js
@@ -11,9 +11,26 @@ const waystones = [
     "waystones:blackstone_brick_waystone",
 ];
 
+const ores_with_missing_tags = {
+    iron: ["archon:cloud_iron"],
+    quartz: [
+        "byg:blue_nether_quartz_ore",
+        "byg:brimstone_nether_quartz_ore",
+    ],
+    lapis: ["betternether:nether_lapis_ore"],
+    redstone: ["betternether:nether_redstone_ore"],
+};
+
 function blockTags(ev) {
     // this allows players to activate other people's waystones
     ev.add("ftbchunks:interact_whitelist", waystones);
+
+    for (let [tag, ores] of Object.entries(ores_with_missing_tags)) {
+        for (ore of ores) {
+            ev.add(`c:${tag}_ores`, ore);
+            ev.add(`c:ores/${tag}`, ore);
+        }
+    }
 }
 
 function addNuggetTags(ev, material, item) {
@@ -25,6 +42,13 @@ function itemTags(ev) {
     // ConsistencyPlus actually doesn't know how tags work
     addNuggetTags(ev, "copper", "consistency_plus:copper_nugget");
     addNuggetTags(ev, "netherite", "consistency_plus:netherite_nugget");
+
+    for (let [tag, ores] of Object.entries(ores_with_missing_tags)) {
+        for (ore of ores) {
+            ev.add(`c:${tag}_ores`, ore);
+            ev.add(`c:ores/${tag}`, ore);
+        }
+    }
 }
 
 onEvent("tags.blocks", blockTags);